+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 - API resource public/get_public_object_count: add filters by object type on taxonomy category
 - Remove docker network as "external" in the general docker compose configuration
 - Zoom the entity map on Europe instead of Luxembourg on oxe-web-admin
-- Database structure: make the "Form.reference" from 20 to 30 chars long
+- Database structure: make the "FormQuestion.reference" from 20 to 30 chars long
+- Database structure: make the "FormQuestion.options" from 500 chars long to TEXT
 - Fix docker image version for API
 
 ### Added